refacto: separate column per file
