+++ frontend/src/features/Dashboard/components/DashboardForm/Columns/style.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+
+export const BaseColumn = styled.div`
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+  scrollbar-gutter: stable;
+  overflow-y: auto;
+
+  padding: 8px;
+`
