+++ frontend/src/features/Dashboard/components/DashboardForm/Columns/utils.ts
@@ -1,0 +1,13 @@
+import type { Dashboard } from '@features/Dashboard/types'
+import type { RefObject } from 'react'
+
+export const scrollToSection = (sectionRef: RefObject<HTMLDivElement>) => {
+  sectionRef.current?.scrollIntoView({ behavior: 'smooth' })
+}
+
+export type ColumnProps = {
+  className: string
+  expandedAccordion: Dashboard.Block | undefined
+  isSelectedAccordionOpen: boolean
+  onExpandedAccordionClick: (type: Dashboard.Block) => void
+}
